requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
Pillow==10.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.2.2
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
urllib3==2.5.0

pytesseract==0.3.10
